requirements.txt
django>=5.0.0
djangorestframework

## iris community package
# db-api
intersystems_iris @ https://github.com/intersystems-community/intersystems-irispython/releases/download/3.7.7/intersystems_iris-3.7.7-py3-none-any.whl
# django-iris
django-iris>=0.2.5b4
# iop - interoperability on python
iris-pex-embedded-python>=2.0.0

#llama_index.embeddings.huggingface
#llama_index.llms.ollama

#for ollama
requests
openai

#for connection to the database
python-dotenv

## for testing
pytest
